refactor(vocabulary): modularize word form into reusable components and hooks
- Extract BasicInfoFields, MetadataFields, ModulesField, RelatedWordsFields, and TopicsField as separate components
- Create useWordForm hook to manage form state and initialization logic
- Create useWordMutation hook to handle API mutations with success and error handling
- Create useTopicAutocomplete hook to manage topic input, filtering, and selection logic with React Query integration
- Simplify WordModal by composing new components and hooks, removing redundant state and effect logic
- Update backend words service to improve code formatting and comment accuracy
- Refine WordSchema pre-save hook and model export formatting
