+++ apps/admin/src/features/vocabulary/components/WordForm/index.ts
@@ -1,0 +1,5 @@
+export { BasicInfoFields } from './BasicInfoFields';
+export { ModulesField } from './ModulesField';
+export { MetadataFields } from './MetadataFields';
+export { RelatedWordsFields } from './RelatedWordsFields';
+export { TopicsField } from './TopicsField';
