+++ apps/backend/src/modules/words/words.service.ts
@@ -28,7 +28,7 @@ export async function resolveTopic(topicInput: string): Promise<string> {
         {
           new: true,
           upsert: true,
-        },
+        }
       );
       return topic._id.toString();
     } catch (error: any) {
@@ -50,7 +50,7 @@ export async function resolveTopic(topicInput: string): Promise<string> {
 // Resolve multiple topics at once
 export async function resolveTopics(topicInputs: string[]): Promise<string[]> {
   const resolvedTopicIds = await Promise.all(
-    topicInputs.map((topicInput) => resolveTopic(topicInput)),
+    topicInputs.map((topicInput) => resolveTopic(topicInput))
   );
 
   // Remove duplicates
@@ -61,7 +61,11 @@ export class WordsService {
   static async create(input: CreateWordInput) {
     const wordData = { ...input };
     // If topics are provided, handle them (they might be names or IDs)
-    if (wordData.topics && Array.isArray(wordData.topics) && wordData.topics.length > 0) {
+    if (
+      wordData.topics &&
+      Array.isArray(wordData.topics) &&
+      wordData.topics.length > 0
+    ) {
       wordData.topics = await resolveTopics(wordData.topics as string[]);
     }
     return Word.create(wordData);
@@ -81,7 +85,10 @@ export class WordsService {
 
     // Search by topic Name
     if (query.topicName) {
-      const escapedTopicName = query.topicName.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+      const escapedTopicName = query.topicName.replace(
+        /[.*+?^${}()|[\]\\]/g,
+        '\\$&'
+      );
       const topics = await Topic.find({
         name: { $regex: escapedTopicName, $options: 'i' },
       }).select('_id');
@@ -99,7 +106,7 @@ export class WordsService {
       }
     }
 
-    // Search across word, meaning, synonyms, and antonyms using indexed searchableText
+    // Search across word, synonyms, and antonyms using indexed searchableText
     if (query.search) {
       const searchLower = query.search.toLowerCase();
       const escapedSearch = searchLower.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
